inline preload functions
