+++ packages/connect-multichain/src/ui/index.ts
@@ -2,16 +2,32 @@
  * Browser/Web UI module entry point
  */
 import { BaseModalFactory } from './ModalFactory';
-import { preload } from './preload';
 import type { FactoryModals } from './modals/types';
 
-export { preload };
+/**
+ * Web-specific preload that loads Stencil custom elements
+ */
+export async function preload(): Promise<void> {
+  if (typeof document === 'undefined') {
+    return;
+  }
+  try {
+    const { defineCustomElements } = await import(
+      '@metamask/multichain-ui/loader'
+    );
+    await defineCustomElements();
+  } catch (error) {
+    console.error('Failed to load customElements:', error);
+  }
+}
 
 /**
  * ModalFactory for browser/web environments.
  * Loads Stencil web components via dynamic import.
  */
-export class ModalFactory<T extends FactoryModals = FactoryModals> extends BaseModalFactory<T> {
+export class ModalFactory<
+  T extends FactoryModals = FactoryModals,
+> extends BaseModalFactory<T> {
   protected async preload(): Promise<void> {
     return preload();
   }
