--- packages/connect-multichain/src/ui/preload.native.ts
@@ -1,6 +1,0 @@
-/**
- * React Native preload - no-op since Stencil web components are not used
- */
-export async function preload(): Promise<void> {
-  // No-op for React Native - web components are not applicable
-}
